requirements.txt
# Minimal requirements for running the API server locally
# (ML dependencies are only needed on Modal)

# Core API dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
requests>=2.31.0

# Supabase & Database
supabase>=2.3.0
python-dotenv>=1.0.0

# Auth libraries
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.1.2

# Rate limiting & security
slowapi>=0.1.9

# Modal (for deployment, not local execution)
modal>=0.69.0

# Configuration
pyyaml>=6.0.0

# Testing dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

